pyproject.toml
[tool.poetry]
name = "pytmux"
version = "0.1.0"
description = ""
authors = ["haoliang <haoliang0.1.2@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]

# linters, formatters for nvim
pylint = "*"
mypy = "*"
flake8 = "*"
isort = "*"
black = "*"
pynvim = "*"

ipython = "^7.24.0"
pytest = "^6.2.4"
memory-profiler = "^0.58.0"

[[tool.poetry.source]]
name = 'tencent_mirror'
url = 'http://mirrors.cloud.tencent.com/pypi/simple'
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
